Refactor satpy.demo subpackage for better organization
